*: enable lfu for analyze
