+++ executor/test/analyzetest/BUILD.bazel
@@ -30,10 +30,8 @@ go_test(
         "//tablecodec",
         "//testkit",
         "//types",
-        "//util",
         "//util/codec",
         "//util/dbterror/exeerrors",
-        "//util/memory",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
